# packages/shared-types/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Iterra API",
    "description": "AI Content Strategy Platform API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/auth/onboarding": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Complete Onboarding",
        "operationId": "complete_onboarding_api_v1_auth_onboarding_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "ittera_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ittera Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OnboardingRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/onboarding": {
      "post": {
        "tags": [
          "onboarding"
        ],
        "summary": "Complete Onboarding",
        "operationId": "complete_onboarding_api_v1_onboarding_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "ittera_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ittera Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OnboardingRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/linkedin/status": {
      "get": {
        "tags": [
          "linkedin"
        ],
        "summary": "Get Linkedin Status",
        "operationId": "get_linkedin_status_api_v1_linkedin_status_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "ittera_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ittera Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LinkedInStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/linkedin/connect/mock": {
      "post": {
        "tags": [
          "linkedin"
        ],
        "summary": "Connect Mock",
        "operationId": "connect_mock_api_v1_linkedin_connect_mock_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "ittera_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ittera Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LinkedInConnectResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/linkedin/sync": {
      "post": {
        "tags": [
          "linkedin"
        ],
        "summary": "Sync",
        "operationId": "sync_api_v1_linkedin_sync_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "ittera_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ittera Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LinkedInSyncResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/content/repurpose": {
      "post": {
        "tags": [
          "content"
        ],
        "summary": "Repurpose",
        "operationId": "repurpose_api_v1_content_repurpose_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "ittera_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ittera Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RepurposeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepurposeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/repurpose/": {
      "post": {
        "tags": [
          "repurpose"
        ],
        "summary": "Repurpose content for other platforms",
        "description": "Maps one idea to multiple platform shaped drafts using `iterra_ai.RepurposeEngine` (EXPERIMENTAL template path in the AI package \u2014 not a production LLM repurposing pipeline yet).",
        "operationId": "repurpose_content_api_v1_repurpose__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "ittera_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ittera Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RepurposeInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RepurposeOutput"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LinkedInConnectResponse": {
        "properties": {
          "connected": {
            "type": "boolean",
            "title": "Connected"
          },
          "platform_username": {
            "type": "string",
            "title": "Platform Username"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "connected",
          "platform_username",
          "message"
        ],
        "title": "LinkedInConnectResponse"
      },
...
      "LinkedInStatusResponse": {
        "properties": {
          "connected": {
            "type": "boolean",
            "title": "Connected"
          },
          "platform_username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Platform Username"
          },
          "last_synced_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Synced At"
          },
          "synced_posts": {
            "type": "integer",
            "title": "Synced Posts",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "connected"
        ],
        "title": "LinkedInStatusResponse"
      },
      "LinkedInSyncResponse": {
        "properties": {
          "synced_posts": {
            "type": "integer",
            "title": "Synced Posts"
          },
          "last_synced_at": {
            "type": "string",
            "format": "date-time",
            "title": "Last Synced At"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "synced_posts",
          "last_synced_at",
          "message"
        ],
        "title": "LinkedInSyncResponse"
      },
...
      "OnboardingRequest": {
        "properties": {
          "full_name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Full Name"
          },
          "niche": {
            "type": "string",
            "maxLength": 160,
            "minLength": 1,
            "title": "Niche"
          },
          "goals": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Goals"
          },
          "primary_platform": {
            "type": "string",
            "enum": [
              "linkedin",
              "instagram",
              "twitter"
            ],
            "title": "Primary Platform",
            "default": "linkedin"
          },
          "storage_preference": {
            "type": "string",
            "enum": [
              "google_drive",
              "local",
              "iterra"
            ],
            "title": "Storage Preference",
            "default": "google_drive"
          }
        },
        "type": "object",
        "required": [
          "full_name",
          "niche"
        ],
        "title": "OnboardingRequest"
      },
...
      "RepurposeInput": {
        "properties": {
          "original_content": {
            "type": "string",
            "title": "Original Content"
          },
          "source_platform": {
            "type": "string",
            "title": "Source Platform"
          },
          "target_platforms": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Target Platforms"
          }
        },
        "type": "object",
        "required": [
          "original_content",
          "source_platform",
          "target_platforms"
        ],
        "title": "RepurposeInput"
      },
      "RepurposeOutput": {
        "properties": {
          "repurposed": {
            "items": {
              "$ref": "#/components/schemas/RepurposedItem"
            },
            "type": "array",
            "title": "Repurposed"
          }
        },
        "type": "object",
        "required": [
          "repurposed"
        ],
        "title": "RepurposeOutput"
      },
      "RepurposeRequest": {
        "properties": {
          "draft_id": {
            "type": "string",
            "title": "Draft Id"
          },
          "target_platform": {
            "type": "string",
            "enum": [
              "instagram",
              "twitter"
            ],
            "title": "Target Platform"
          }
        },
        "type": "object",
        "required": [
          "draft_id",
          "target_platform"
        ],
        "title": "RepurposeRequest"
      },
      "RepurposeResponse": {
        "properties": {
          "draft_id": {
            "type": "string",
            "title": "Draft Id"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "platform": {
            "type": "string",
            "title": "Platform"
          }
        },
        "type": "object",
        "required": [
          "draft_id",
          "content",
          "platform"
        ],
        "title": "RepurposeResponse"
      },
      "RepurposedItem": {
        "properties": {
          "platform": {
            "type": "string",
            "title": "Platform"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "format": {
            "type": "string",
            "title": "Format"
          }
        },
        "type": "object",
        "required": [
          "platform",
          "content",
          "format"
        ],
        "title": "RepurposedItem"
      },
...
      "UserResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "full_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "niche": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Niche"
          },
          "goals": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Goals"
          },
          "primary_platform": {
            "type": "string",
            "title": "Primary Platform",
            "default": "linkedin"
          },
          "onboarding_complete": {
            "type": "boolean",
            "title": "Onboarding Complete",
            "default": false
          },
          "storage_preference": {
            "type": "string",
            "title": "Storage Preference",
            "default": "google_drive"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "name",
          "is_active"
        ],
        "title": "UserResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}